Break PropelIndex out into its own file.
